+++ src/main/java/org/xodium/illyriacore/interfaces/SI.java
@@ -9,11 +9,8 @@ import net.kyori.adventure.sound.Sound;
  * The Sound Interface.
  */
 public interface SI {
-    /**
-     * The sound played when the immunity bar is activated.
-     */
     @NotNull
-    Sound IMMUNITY_BAR_SOUND = Sound.sound(
+    Sound IMMUNITY_BAR = Sound.sound(
             Key.key("minecraft", "block.note_block.pling"),
             Sound.Source.MASTER, 1f, 1f);
 }
